Move mempool interface to a separate file
